appointment-service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
requests==2.31.0
pymysql==1.1.0
cryptography==41.0.7
PyJWT==2.8.0
psutil==5.9.6

